api/test.py
import os
from google import genai
from dotenv import load_dotenv

load_dotenv()
client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

print("Listing all available models for your key:")
try:
    for m in client.models.list():
        # This will print everything available
        print(f"- {m.name}")
except Exception as e:
    print(f"Error: {e}")