api/requirements.txt
fastapi
google-genai
motor
bcrypt
pymupdf
python-dotenv